refactor : 파일 분리, 메소드 이름 수정
