+++ iOS/baseball-game/baseball-game/Selection/Model/GameInfo.swift
@@ -1,0 +1,14 @@
+//
+//  GameInfo.swift
+//  baseball-game
+//
+//  Created by Song on 2021/05/07.
+//
+
+import Foundation
+
+struct GameInfo: Codable {
+    var userID: String = ""
+    var gameID: Int = 0
+    var team: String = ""
+}
