+++ iOS/baseball-game/baseball-game/Selection/Model/Team.swift
@@ -1,0 +1,13 @@
+//
+//  Team.swift
+//  baseball-game
+//
+//  Created by Song on 2021/05/07.
+//
+
+import Foundation
+
+struct Team: Codable, Hashable {
+    var team: String = ""
+    var status: String = ""
+}
